Injeta dependências no projeto forms e refatora filtros do serviços

+++ Cod3rsGrowth.Infra/migrations/HabilidadeMigration.cs
@@ -8,7 +8,7 @@ namespace Cod3rsGrowth.Infra.Migrations
         public override void Up()
         {
             Create.Table("habilidades")
-                .WithColumn("id").AsInt64().PrimaryKey().Identity()
+                .WithColumn("id").AsInt32().PrimaryKey().Identity()
                 .WithColumn("nome").AsString(50).NotNullable()
                 .WithColumn("descricao").AsString(200).Nullable()
                 .WithColumn("criado_em").AsDateTime().Nullable().WithDefault(SystemMethods.CurrentDateTime)
